Move remaining mozorg app views (all nonloacle views) into base app
