+++ bedrock/base/templates/base/gpc.json
