+++ bedrock/base/templates/base/locales.html
