+++ bedrock/base/tests/test_views.py
@@ -5,10 +5,13 @@
 from unittest.mock import patch
 
 from django.conf import settings
-from django.test import RequestFactory, TestCase
+from django.test import TestCase as DjangoTestCase
+from django.test.client import RequestFactory
 
 import pytest
 
+from bedrock.base import views
+from bedrock.base.tests import TestCase
 from bedrock.base.views import GeoTemplateView
 
 geo_template_view = GeoTemplateView.as_view(
@@ -20,7 +23,7 @@ geo_template_view = GeoTemplateView.as_view(
 )
 
 
-class TestGeoTemplateView(TestCase):
+class TestGeoTemplateView(DjangoTestCase):
     def get_template(self, country):
         with patch("bedrock.firefox.views.l10n_utils.render") as render_mock:
             with patch("bedrock.base.views.get_country_from_request") as geo_mock:
@@ -46,3 +49,35 @@ class TestGeoTemplateView(TestCase):
 @pytest.mark.django_db
 def test_csrf_view_is_custom_one():
     assert settings.CSRF_FAILURE_VIEW == "bedrock.base.views.csrf_failure"
+
+
+class TestRobots(TestCase):
+    def setUp(self):
+        self.rf = RequestFactory()
+        self.view = views.Robots()
+
+    def test_production_disallow_all_is_false(self):
+        self.view.request = self.rf.get("/", HTTP_HOST="www.mozilla.org")
+        self.assertFalse(self.view.get_context_data()["disallow_all"])
+
+    def test_non_production_disallow_all_is_true(self):
+        self.view.request = self.rf.get("/", HTTP_HOST="www.allizom.org")
+        self.assertTrue(self.view.get_context_data()["disallow_all"])
+
+    def test_robots_no_redirect(self):
+        response = self.client.get("/robots.txt", headers={"host": "www.mozilla.org"})
+        self.assertEqual(response.status_code, 200)
+        self.assertFalse(response.context_data["disallow_all"])
+        self.assertEqual(response.get("Content-Type"), "text/plain")
+
+
+class TestSecurityDotTxt(TestCase):
+    def setUp(self):
+        self.rf = RequestFactory()
+        self.view = views.SecurityDotTxt()
+
+    def test_no_redirect(self):
+        response = self.client.get("/.well-known/security.txt", headers={"host": "www.mozilla.org"})
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.get("Content-Type"), "text/plain")
+        self.assertContains(response, "security@mozilla.org")
