+++ bedrock/base/views.py
@@ -12,13 +12,16 @@ from django.conf import settings
 from django.shortcuts import render
 from django.views.decorators.cache import never_cache
 from django.views.decorators.http import require_safe
+from django.views.generic import TemplateView
 
 import timeago
+from product_details import product_details
 from waffle.models import Switch
 
 from bedrock.base.geo import get_country_from_request
 from bedrock.utils import git
 from lib import l10n_utils
+from lib.l10n_utils import RequireSafeMixin
 
 
 class GeoTemplateView(l10n_utils.L10nTemplateView):
@@ -172,3 +175,32 @@ def page_not_found_view(request, exception=None, template_name="404.html"):
 
 def csrf_failure(request, reason="CSRF failure", template_name="403_csrf.html"):
     return render(request, template_name, status=403)
+
+
+class Robots(RequireSafeMixin, TemplateView):
+    template_name = "base/robots.txt"
+    content_type = "text/plain"
+
+    def get_context_data(self, **kwargs):
+        hostname = self.request.get_host()
+        return {"disallow_all": not hostname == "www.mozilla.org"}
+
+
+class SecurityDotTxt(RequireSafeMixin, TemplateView):
+    # https://github.com/mozilla/bedrock/issues/14173
+    # served under .well-known/security.txt
+    template_name = "base/security.txt"
+    content_type = "text/plain"
+
+
+class GpcDotJson(RequireSafeMixin, TemplateView):
+    # https://github.com/mozilla/bedrock/issues/14213
+    # served under .well-known/gpc.json
+    template_name = "base/gpc.json"
+    content_type = "application/json"
+
+
+@require_safe
+def locales(request):
+    context = {"languages": product_details.languages}
+    return l10n_utils.render(request, "base/locales.html", context)
