+++ bedrock/urls.py
@@ -32,7 +32,7 @@ urlpatterns = bedrock_i18n_patterns(
 
 # Paths that must not have a locale prefix
 urlpatterns += (
-    path("", include("bedrock.mozorg.nonlocale_urls")),
+    path("", include("bedrock.base.nonlocale_urls")),
     path("", include("bedrock.sitemaps.urls")),
     path("healthz/", watchman_views.ping, name="watchman.ping"),
     path("readiness/", watchman_views.status, name="watchman.status"),
